Tools for use in the Malware unpacking exercise
None of these tools are perfect, some will crash if you use them outside their defined parameters. They are not commercial products and should not be used in production. I use them myself to unpack malware, they work fine for that purpose so.

Sourcecode will be released later on when i feel like it.