Tools to be used for the Malware unpacking exercise
An assortment of tools to be used for the exercise, very beta and should not be used in production. I use them myself to unpack malware and they works for me. Source code to be released later on when i feel like it. Description on how to use the files (Tools-Description.txt) are in the zip file.